requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0
pyserial==3.5
PyYAML==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
